Refactoring based on cogs
* Moving the commands separate files
