+++ comandos/ping.py
@@ -1,0 +1,9 @@
+from discord.ext import commands
+
+class Ping(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
+
+    @commands.command(name="ping", help="Comando de prueba", pass_context=True)
+    async def pingpong(self, ctx):
+        await ctx.send("pong")
